tests/wpt/web-platform-tests/navigation-api/updateCurrentEntry-method/exception-order-initial-about-blank-unserializablestate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe"></iframe>

<script>
test(() => {
  assert_throws_dom("InvalidStateError", iframe.contentWindow.DOMException, () => {
    iframe.contentWindow.navigation.updateCurrentEntry({ state: document.body });
  });
  assert_equals(navigation.currentEntry.getState(), undefined);
}, `updateCurrentEntry() with unserializable state on the initial about:blank must throw an "InvalidStateError", not a "DataCloneError"`);
</script>
